requirements.txt
discord.py==2.0.1
Flask==2.1.2
python-dotenv==0.21.0
pytz==2022.1
